requirements.txt
PyYAML
check-jsonschema
